Revert "Merge remote-tracking branch 'origin/release_2.0' into release_2.1"
This reverts commit fc476bcad7284ba3a225143aaf4b5792c14c37be, reversing
changes made to 9c0df5435f145c9e9e7d48c2cca9b648458714b6.
