+++ src/main/java/gwt/material/design/jquery/client/api/JQuery.java
@@ -39,10 +39,10 @@ import com.google.gwt.core.client.ScriptInjector;
 import com.google.gwt.dom.client.Element;
 import com.google.gwt.dom.client.Node;
 import com.google.gwt.user.client.ui.UIObject;
-import gwt.material.design.jscore.client.api.Functions.Func;
-import gwt.material.design.jscore.client.api.Functions.Func2;
-import gwt.material.design.jscore.client.api.Functions.Func3;
-import gwt.material.design.jscore.client.api.Functions.FuncRet2;
+import gwt.material.design.jquery.client.api.Functions.Func;
+import gwt.material.design.jquery.client.api.Functions.Func2;
+import gwt.material.design.jquery.client.api.Functions.Func3;
+import gwt.material.design.jquery.client.api.Functions.FuncRet2;
 import jsinterop.annotations.JsMethod;
 import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsPackage;
