+++ src/main/java/gwt/material/design/jquery/client/api/JQueryElement.java
@@ -37,20 +37,19 @@ package gwt.material.design.jquery.client.api;
 
 
 import com.google.gwt.dom.client.Element;
-import gwt.material.design.jscore.client.api.Event;
-import gwt.material.design.jscore.client.api.Functions.EventFunc;
-import gwt.material.design.jscore.client.api.Functions.EventFunc1;
-import gwt.material.design.jscore.client.api.Functions.EventFunc2;
-import gwt.material.design.jscore.client.api.Functions.EventFunc3;
-import gwt.material.design.jscore.client.api.Functions.Func;
-import gwt.material.design.jscore.client.api.Functions.Func1;
-import gwt.material.design.jscore.client.api.Functions.Func2;
-import gwt.material.design.jscore.client.api.Functions.Func3;
-import gwt.material.design.jscore.client.api.Functions.FuncRet1;
-import gwt.material.design.jscore.client.api.Functions.FuncRet2;
-import gwt.material.design.jscore.client.api.Functions.FuncRet3;
-import gwt.material.design.jscore.client.api.Functions.KeyEventFunc;
-import gwt.material.design.jscore.client.api.Functions.MouseEventFunc;
+import gwt.material.design.jquery.client.api.Functions.EventFunc;
+import gwt.material.design.jquery.client.api.Functions.EventFunc1;
+import gwt.material.design.jquery.client.api.Functions.EventFunc2;
+import gwt.material.design.jquery.client.api.Functions.EventFunc3;
+import gwt.material.design.jquery.client.api.Functions.Func;
+import gwt.material.design.jquery.client.api.Functions.Func1;
+import gwt.material.design.jquery.client.api.Functions.Func2;
+import gwt.material.design.jquery.client.api.Functions.Func3;
+import gwt.material.design.jquery.client.api.Functions.FuncRet1;
+import gwt.material.design.jquery.client.api.Functions.FuncRet2;
+import gwt.material.design.jquery.client.api.Functions.FuncRet3;
+import gwt.material.design.jquery.client.api.Functions.KeyEventFunc;
+import gwt.material.design.jquery.client.api.Functions.MouseEventFunc;
 import gwt.material.design.jscore.client.api.core.Node;
 import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsProperty;
