+++ src/main/java/gwt/material/design/jscore/client/api/file/DirectoryEntry.java
@@ -1,6 +1,6 @@
 package gwt.material.design.jscore.client.api.file;
 
-import gwt.material.design.jscore.client.api.Functions;
+import gwt.material.design.jquery.client.api.Functions;
 import gwt.material.design.jscore.client.api.JsObject;
 import jsinterop.annotations.JsMethod;
 import jsinterop.annotations.JsPackage;
